% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{brf_describe}
\alias{brf_describe}
\title{Get a briefing of a data frame}
\usage{
brf_describe(
  df,
  groupings = NULL,
  na.rm = F,
  output_filename = "briefing.html",
  output_dir = NULL
)
}
\arguments{
\item{df}{the data frame to summarise}

\item{groupings}{The columns to use as grouping variables as a character
vector.}

\item{output_filename}{The name of the .html report (default: "briefing.html")}

\item{output_dir}{The folder where to put the report. (default: `NULL`)
If set to `NULL` the report will be put in the current working directory
Otherwise this parameter will be passed to the `output_dir` parameter of rmarkdown::render()}
}
\description{
Get a briefing of a data frame
}
