% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatted.R
\name{brf_formatted_cat}
\alias{brf_formatted_cat}
\title{Format a categorical column summary with `kable()`}
\usage{
brf_formatted_cat(df, data_col, grouping_col = NULL, decimals = 1)
}
\arguments{
\item{df}{a data frame}

\item{data_col}{the unquoted name of the column to summarise}

\item{grouping_col}{the unquoted name of the column to use for groupings}
}
\value{
the summary as a `kable table`
}
\description{
Format a categorical column summary with `kable()`
}
