% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatted.R
\name{brf_formatted_cat_lvl}
\alias{brf_formatted_cat_lvl}
\title{Format a categorical column levels summary with `gt()`}
\usage{
brf_formatted_cat_lvl(df, data_col, grouping_col = NULL, na.rm = F,
  decimals = 1)
}
\arguments{
\item{df}{a data frame}

\item{data_col}{the unquoted name of the column to summarise}

\item{grouping_col}{the unquoted name of the column to use for groupings}

\item{na.rm}{should `NA`s be removed (default: FALSE)}
}
\value{
the summary as a `kable table`
}
\description{
Format a categorical column levels summary with `gt()`
}
