% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatted.R
\name{brf_formatted_cat_lvl_wide}
\alias{brf_formatted_cat_lvl_wide}
\title{Format a categorical column levels summary with `gt()` - wide version}
\usage{
brf_formatted_cat_lvl_wide(df, data_col, grouping_col, na.rm = F)
}
\arguments{
\item{df}{a data frame}

\item{data_col}{the unquoted name of the column to summarise}

\item{grouping_col}{the unquoted name of the column to use for groupings}

\item{na.rm}{should `NA`s be removed (default: FALSE)}
}
\value{
the summary as a `gt_tbl`
}
\description{
Format a categorical column levels summary with `gt()` - wide version
}
