% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatted.R
\name{brf_formatted_df}
\alias{brf_formatted_df}
\title{Format a data frame summary with `gt()`}
\usage{
brf_formatted_df(df)
}
\arguments{
\item{df}{the data frame to summarise}
}
\value{
the summary as a `gt_tbl`
}
\description{
Format a data frame summary with `gt()`
}
