% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatted.R
\name{brf_formatted_num}
\alias{brf_formatted_num}
\title{Format a numerical column summary with `gt()`}
\usage{
brf_formatted_num(df, data_col, grouping_col = NULL, decimals = 1)
}
\arguments{
\item{df}{a data frame}

\item{data_col}{the unquoted name of the column to summarise}

\item{grouping_col}{the unquoted name of the column to use for groupings}
}
\value{
the summary as a `gt_tbl`
}
\description{
Format a numerical column summary with `gt()`
}
