% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{brf_plot_cat_count}
\alias{brf_plot_cat_count}
\title{Make a plot of a categorical column in a data frame, count barplot}
\usage{
brf_plot_cat_count(df, data_col, grouping_col = NULL, na.rm = FALSE)
}
\arguments{
\item{df}{a data frame}

\item{data_col}{the unquoted name of the column to summarise}

\item{grouping_col}{the unquoted name of the column to use for groupings}

\item{na.rm}{should `NA`s be removed (default: FALSE)}
}
\value{
a ggplot object
}
\description{
Make a plot of a categorical column in a data frame, count barplot
}
