% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{brf_plot_num_violin}
\alias{brf_plot_num_violin}
\title{Make a violin plot of a numerical column in a data frame}
\usage{
brf_plot_num_violin(df, num_col, grouping_col = NULL, na.rm = FALSE)
}
\arguments{
\item{df}{a data frame}

\item{grouping_col}{the unquoted name of the column to use for groupings}

\item{na.rm}{should `NA`s be removed (default: FALSE)}
}
\value{
a ggplot object
}
\description{
Make a violin plot of a numerical column in a data frame
}
