% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd.R
\name{brf_rmd_df}
\alias{brf_rmd_df}
\title{Outputs an RMarkDown formatted chunk for a data frame summary}
\usage{
brf_rmd_df(df)
}
\arguments{
\item{df}{the data frame to summarise}
}
\description{
Outputs an RMarkDown formatted chunk for a data frame summary
}
