% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd.R
\name{brf_rmd_dispatch}
\alias{brf_rmd_dispatch}
\title{Choose the correct RMarkDown function depending on the column type}
\usage{
brf_rmd_dispatch(df, data_col, grouping_col = NULL, na.rm = F)
}
\arguments{
\item{df}{a data frame}

\item{data_col}{the unquoted name of the column to summarise}

\item{grouping_col}{the unquoted name of the column to use for groupings}

\item{na.rm}{should `NA`s be removed (default: FALSE)}
}
\description{
Choose the correct RMarkDown function depending on the column type
}
