% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd.R
\name{brf_rmd_num}
\alias{brf_rmd_num}
\title{Outputs an RMarkDown formatted chunk for a numerical column}
\usage{
brf_rmd_num(df, data_col, grouping_col = NULL, na.rm = F)
}
\arguments{
\item{df}{a data frame}

\item{data_col}{the unquoted name of the column to summarise}

\item{grouping_col}{the unquoted name of the column to use for groupings}

\item{na.rm}{should `NA`s be removed (default: FALSE)}
}
\description{
Outputs an RMarkDown formatted chunk for a numerical column
}
