% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{brf_summary_cat_lvl}
\alias{brf_summary_cat_lvl}
\title{Summarise the levels of a categorical column in a data frame with a grouping option}
\usage{
brf_summary_cat_lvl(df, data_col, grouping_col = NULL, na.rm = F)
}
\arguments{
\item{df}{a data frame}

\item{data_col}{the unquoted name of the column to summarise}

\item{grouping_col}{the unquoted name of the column to use for groupings}
}
\value{
a dataframe containing the summary informations
}
\description{
Summarise the levels of a categorical column in a data frame with a grouping option
}
