% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{brf_summary_df}
\alias{brf_summary_df}
\title{Summarise a data frame}
\usage{
brf_summary_df(df)
}
\arguments{
\item{df}{the data frame to summarise}
}
\value{
a one-row data with summarising df
}
\description{
Produce the summary of a data frame.
}
