% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{dflt_kable}
\alias{dflt_kable}
\title{Defautl table format for my summaries}
\usage{
dflt_kable(df, grouping_col = NULL)
}
\arguments{
\item{df}{the data frame to make into a table}

\item{grouping_col}{the grouping variable if any}
}
\value{
a formatted kable table
}
\description{
Defautl table format for my summaries
}
