% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{factor_infreq}
\alias{factor_infreq}
\title{Reassign levels in a factor using frequency}
\usage{
factor_infreq(x)
}
\arguments{
\item{x}{a factor}
}
\value{
a factor ordered by decreasing frequencies of its levels
}
\description{
Reassign levels in a factor using frequency
}
