% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{brf_describe}
\alias{brf_describe}
\title{Get a briefing of a data frame}
\usage{
brf_describe(df, groupings = NULL, na.rm = F,
  output_filename = "briefing.html", output_dir = NULL)
}
\description{
Get a briefing of a data frame
}
