% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/briefr.R
\docType{package}
\name{briefr-package}
\alias{briefr}
\alias{briefr-package}
\title{briefr: Get a briefing of your data frame}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
